get_stock_data/requirements.txt
yfinance>=0.2.0
pandas>=1.3.0
pymysql>=1.1.2
python-dotenv>=1.1.1
